refactor disc compiler
