+++ torch_disc/torch_disc/csrc/disc_compiler/passes/disc_class.h
