+++ torch_disc/torch_disc/csrc/disc_compiler/passes/io.h
@@ -9,6 +9,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <sys/stat.h>
+#include <unistd.h>
+
 #include <sstream>
 
 namespace torch_disc {
