+++ torch_disc/torch_disc/csrc/disc_compiler/passes/register_disc_class.h
@@ -1,0 +1,22 @@
+// Copyright 2022 The BladeDISC Authors. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#pragma once
+
+#include <torch/script.h>
+
+namespace torch_disc {
+namespace compiler {
+
+std::vector<c10::IValue> RegisterDiscClass(
+    const std::shared_ptr<torch::jit::Graph>& graph);
+
+}  //  namespace compiler
+}  //  namespace torch_disc
